[logs] add preview files listing
